add test project for mix using shared library and static library
libtest_1.so, /static_lib/lib/ics/libstagefright/libtest_1.so
libtest_2.so, /static_lib/lib/ics/libutils/libtest_2.so

libtest_static_lib.a, /static_lib/libtest_static_lib.a, it is linked
with libtest_1.so and libtest_2.so.

libtest_shared_lib.so, test/test_shared_lib.so, it is linked with
libtest_static_lib.a with whole-archive enabled.